requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.29,<1.0
pydantic>=2.6,<3.0
PyYAML>=6.0,<7.0
streamlit>=1.32,<2.0
pandas>=2.2,<3.0
requests>=2.31,<3.0


